requirements.txt
-i https://pypi.org/simple
aiofiles==0.5.0
aiohttp==3.6.2
aiomysql==0.0.20
alembic==1.4.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
discord-py==1.3.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.4.28
httpx==0.12.1
hyperframe==5.2.0
idna==2.9
mako==1.1.2
markupsafe==1.1.1
multidict==4.7.5
pillow==7.1.2
pycparser==2.20
pymysql==0.9.2
python-dateutil==2.8.1
python-editor==1.0.4
python-i18n[yaml]==0.3.7
pyyaml==5.3.1
rfc3986==1.4.0
six==1.14.0
sniffio==1.1.0
sqlalchemy==1.3.16
tabulate==0.8.7
urllib3==1.25.9
websockets==8.1
yarl==1.4.2
